Add show next proof
Showing the raw encoded proofs is doesn't mean much and presents UI
challenges. We can instead let the user walk the proof chain to inspect
each subsequesnt proof.

This would be even better if the user could go backwards, but for now
they can only go forward through the proofs.

The example is updated with a UCAN that shows how this works as well.