Fix collapsing textarea and mobile layouts
The encode textarea would collapse horizontally with some values. We
want it to keep the same width as the decoded section.
Changing to a row and column oriented layouts was incosistent and the
grid helps us here. Also, the Explanation section would force a wide
layout on smaller devices because of the long did strings (that can't
have a scrollbar). Removing the value column on smaller devices fixes
that.