Add proofs validation
The current version of the spec calls for an array of proofs instead of
a single proof. We want to validate each proof to check that all proof
audiences match the issuer and all proofs are valid.
The examples and show next proof are updated to account for this, though
show next proof will only show the first proof. We will show multiple
proofs coming up.
The tests were also updated and new test cases were added.