Add proof explorer (#9)
Add explorer to inspect any proof. The explorer shows the selected UCAN,
delegates, and proofs. Each node also has an absolute label to make
it easier to track them.
Hints show issuer, audience, and whether or not the UCAN is valid.
A triangle in the corner indicates UCANs that are invalid. The triangle
is faded out when the UCAN cannot be decoded.